refactor: replace barrel imports with direct source imports and fix broken test mocks
