+++ src/components/ActiveMatchScreen/ActiveMatchScreen.tsx
@@ -2,12 +2,14 @@ import { useCallback, useEffect, useMemo, useState } from 'react'
 import { useNavigate } from '@tanstack/react-router'
 import { useTranslation } from 'react-i18next'
 
-import { Layout } from '@/components/Layout'
+import { Layout } from '@/components/Layout/Layout'
 import { TeamPanel } from './TeamPanel'
 import { SetsCard } from './SetsCard'
 import { InfoCard } from './InfoCard'
 import { SideSwitchPrompt } from './SideSwitchPrompt'
-import { Button, Chip, TopBar } from '@/components/ui'
+import { Button } from '@/components/ui/Button'
+import { Chip } from '@/components/ui/Chip'
+import { TopBar } from '@/components/ui/TopBar'
 import { useMatchTimer } from './useMatchTimer'
 import { useMatchSession } from './useMatchSession'
