+++ src/components/ActiveMatchScreen/SetsCard/SetsCard.tsx
@@ -1,7 +1,7 @@
 import { useTranslation } from 'react-i18next'
 
 import type { MatchSetState } from '@/core/match'
-import { Card } from '@/components/ui'
+import { Card } from '@/components/ui/Card'
 
 import styles from './SetsCard.module.css'
